refactor(#843): Infra Pulumi src segmentation skeleton
Split infra/index.ts into src modules (config, buckets, web-routes,
hardening, staging, outputs). Thin composition root keeps public exports
stable for topology tests. ESC/lifecycle TODOs for #674/#521.

Closes #843
Parent #829
