+++ infra/index.ts
@@ -1,458 +1,18 @@
-import * as pulumi from "@pulumi/pulumi";
-import * as cloudflare from "@pulumi/cloudflare";
+// Composition root for Cloudflare Pulumi IaC.
+// Resource modules register side effects on import; this file re-exports the
+// stable public surface topology tests and stack consumers rely on.
+// Layout: see src/README.md
 
-// Wave 0 spike (2026-06-23) validated:
-//   - Pulumi auto-provisions Cloudflare resources (R2 bucket + DNS + Workers
-//     routes) with only a scoped CLOUDFLARE_API_TOKEN, using local state.
-//   - `wrangler deploy` of a no-route worker does NOT clobber Pulumi-managed
-//     routes → "routes belong to Pulumi, worker code belongs to wrangler".
-//
-// Wave 2 (Task 4): catalog infra added below.
-// Wave 2+ will declare real infra parameterized per stack (prod / staging) —
-// R2 (media + Pulumi state), Worker routes/custom domains, DNS, secrets.
-// See docs/superpowers/specs/2026-06-23-platform-monorepo-cf-deploy-design.md
+import "./src/buckets.ts";
+import "./src/web-routes.ts";
+import "./src/hardening.ts";
+import "./src/staging.ts";
 
-const config = new pulumi.Config();
-const stack = pulumi.getStack();
-const mediaBucketName = stack === "prod" ? "catalog-media" : `catalog-media-${stack}`;
-const mapTilesBucketName =
-  stack === "prod" ? "map-tiles" : stack === "staging" ? "map-tiles-staging" : `map-tiles-${stack}`;
-const accountId = config.require("cloudflareAccountId");
-const webRoutesEnabled = config.getBoolean("webRoutesEnabled") ?? false;
-
-// Custom Domains own the web origins because they provide the originless Worker
-// hostname and certificate; explicit routes then take precedence for the edge
-// API paths. One flag gates DNS and those narrowed routes together, so enabling
-// it cannot publish the apex while it still falls through to the edge JSON 404.
-// Keep this false by default; enabling requires:
-//   pulumi config set webRoutesEnabled true
-//   pulumi config set cloudflareZoneId <zone id>
-//   pulumi config set webDomain <domain>
-//   pulumi config set stagingDomain staging.animichi.com
-//   pulumi config set wwwDomain www.animichi.com
-
-// ── Catalog: Neon DATABASE_URL (managed secret — stored in Pulumi config) ────
-// Optional and operator-set per stack via:
-//   neonctl connection-string main --project-id $NEON_PROJECT_ID
-//   pulumi config set --secret catalogDatabaseUrl <connstr> --stack prod
-// No Hyperdrive needed: catalog uses @neondatabase/serverless (neon-http,
-// HTTP transport, no raw socket). The CI deploy-catalog job passes this value
-// to `wrangler secret put DATABASE_URL` (see .github/workflows/ci.yml).
-// @pulumi/cloudflare v6 does not expose a WorkersSecret resource; the secret
-// is passed through CI environment using the Pulumi output below.
-export const catalogDatabaseUrl = config.getSecret("catalogDatabaseUrl");
-
-// ── Catalog: R2 media bucket ──────────────────────────────────────────────────
-// catalog Worker uses MEDIA_BUCKET (see workers/catalog/src/media/r2.ts) for
-// lazy-cached pilgrimage point photos. Pulumi owns the bucket; wrangler.toml
-// references it by name (bucket_name = "catalog-media").
-// catalog has NO public route — it is a service-binding target from edge Worker.
-// #487: deleteBeforeReplace avoids account-global name collisions during replacement.
-// protect prevents accidental deletion of unrecoverable bucket data.
-const catalogMediaBucket = new cloudflare.R2Bucket("catalog-media", {
-  accountId,
-  name: mediaBucketName,
-  location: "apac",
-}, { protect: true, deleteBeforeReplace: true });
-
-// Map tiles, glyphs, sprites, and style JSON are private R2 objects. The edge
-// Worker is the sole public reader through `/tiles/*`; there is intentionally
-// no R2 public bucket/domain configuration here. Production and staging use
-// stable names consumed by their matching wrangler environment; an unrecognised
-// preview stack gets an isolated suffix instead of sharing either live bucket.
-const mapTilesBucket = new cloudflare.R2Bucket("map-tiles", {
-  accountId,
-  name: mapTilesBucketName,
-  location: "apac",
-}, { protect: true, deleteBeforeReplace: true });
-
-if (webRoutesEnabled) {
-  const cloudflareZoneId = config.require("cloudflareZoneId");
-  const webScript = stack === "prod" ? "animichi-web" : `animichi-web-${stack}`;
-  const edgeScript = stack === "prod" ? "animichi" : `animichi-${stack}`;
-
-  // The hostname this stack serves: prod owns the apex, every other stack its
-  // own subdomain. Both get the SAME split — the Custom Domain is the origin
-  // for pages, and the four routes run ahead of it for the API/map surfaces.
-  // Staging must not be an exception: `apps/web` calls `/v1/chat`,
-  // `/v1/photo-search`, `/v1/conversations/...` relative to its own origin, so a
-  // staging hostname pointed wholly at the web Worker has no chat at all.
-  // Only two stacks have a hostname today. A third would fall into the
-  // non-prod branch and read `stagingDomain` — and the natural way to bootstrap
-  // one is copying `Pulumi.staging.yaml`, which carries `staging.animichi.com`.
-  // That stack would then claim staging's hostname with its own Workers and
-  // silently take it over on the next `pulumi up`. Fail loudly instead; adding
-  // a stack should be a deliberate edit here, not an inherited surprise.
-  if (stack !== "prod" && stack !== "staging") {
-    throw new Error(
-      `stack "${stack}" has no hostname mapping. Add one here rather than ` +
-        `letting it inherit stagingDomain — see the comment above.`,
-    );
-  }
-  const apexDomain =
-    stack === "prod" ? config.require("webDomain") : config.require("stagingDomain");
-
-  new cloudflare.WorkersCustomDomain("animichi-web-domain", {
-    accountId,
-    hostname: apexDomain,
-    service: webScript,
-    zoneId: cloudflareZoneId,
-  }, { deleteBeforeReplace: true });
-
-  new cloudflare.WorkersRoute("animichi-edge-v1-route", {
-    zoneId: cloudflareZoneId,
-    pattern: `${apexDomain}/v1/*`,
-    script: edgeScript,
-  }, { deleteBeforeReplace: true });
-
-  new cloudflare.WorkersRoute("animichi-edge-img-route", {
-    zoneId: cloudflareZoneId,
-    pattern: `${apexDomain}/img/*`,
-    script: edgeScript,
-  }, { deleteBeforeReplace: true });
-
-  new cloudflare.WorkersRoute("animichi-edge-tiles-route", {
-    zoneId: cloudflareZoneId,
-    pattern: `${apexDomain}/tiles/*`,
-    script: edgeScript,
-  }, { deleteBeforeReplace: true });
-
-  new cloudflare.WorkersRoute("animichi-edge-healthz-route", {
-    zoneId: cloudflareZoneId,
-    pattern: `${apexDomain}/healthz`,
-    script: edgeScript,
-  }, { deleteBeforeReplace: true });
-
-  // Legacy-domain 301s (seo-geo-plan §3 item 2 / iter-0 AC): the retired
-  // production domains (`seichijunrei.app`, `seichijunrei.zhenjia.dev`; and
-  // `aninavi.app` if held) 301 onto the canonical apex with path + query
-  // preserved. Each legacy domain must be its own Cloudflare zone (DNS
-  // delegated to CF — a manual-ops step, #545), so the config lists *zone
-  // ids*, not domains: a rule in the animichi.com zone cannot match a
-  // hostname another zone owns. Absent/empty config is a deliberate no-op
-  // until an owner onboards a legacy domain; the ruleset appears in the
-  // legacy zone only, and `expression: "true"` matches every request there
-  // because that zone exists solely to redirect.
-  if (stack === "prod") {
-    const legacyZoneIds = config.getObject<string[]>("legacyRedirectZones") ?? [];
-    validateLegacyRedirectZones(legacyZoneIds);
-    for (const legacyZoneId of legacyZoneIds) {
-      new cloudflare.Ruleset(`animichi-legacy-redirect-${legacyZoneId}`, {
-        zoneId: legacyZoneId,
-        name: `animichi legacy redirect ${legacyZoneId}`,
-        kind: "zone",
-        phase: "http_request_dynamic_redirect",
-        description: "301 the legacy domain onto the canonical apex.",
-        rules: [
-          {
-            action: "redirect",
-            expression: "true",
-            description: "301 every path onto the canonical apex path.",
-            enabled: true,
-            actionParameters: {
-              fromValue: {
-                statusCode: 301,
-                preserveQueryString: true,
-                targetUrl: {
-                  expression: `concat("https://${apexDomain}", http.request.uri.path)`,
-                },
-              },
-            },
-          },
-        ],
-      });
-    }
-  }
-
-  // `www` is prod-only: there is no `www.staging.animichi.com`.
-  if (stack === "prod") {
-    const wwwDomain = config.require("wwwDomain");
-
-    new cloudflare.DnsRecord("animichi-www-placeholder", {
-      zoneId: cloudflareZoneId,
-      name: wwwDomain,
-      type: "A",
-      content: "192.0.2.0",
-      proxied: true,
-      ttl: 1,
-    });
-
-    new cloudflare.Ruleset("animichi-www-redirect", {
-      zoneId: cloudflareZoneId,
-      name: "www to apex redirect",
-      kind: "zone",
-      phase: "http_request_dynamic_redirect",
-      description: "Redirect the placeholder www hostname to the apex.",
-      rules: [
-        {
-          action: "redirect",
-          expression: `http.host eq "${wwwDomain}"`,
-          description: "Redirect www traffic to the apex hostname.",
-          enabled: true,
-          actionParameters: {
-            fromValue: {
-              statusCode: 301,
-              preserveQueryString: true,
-              targetUrl: {
-                expression: `concat("https://${apexDomain}", http.request.uri.path)`,
-              },
-            },
-          },
-      },
-    ],
-  });
-  }
-
-}
-
-// ── Zone hardening: DNSSEC, CAA, API rate limit, HSTS ───────────────────────
-// Owner ruling 2026-08-05 (prod security review), restructured per PR #776:
-// these four are zone-scoped, and prod is the single owner of zone metadata.
-// They started inside the `webRoutesEnabled` block, which meant staging and
-// prod would BOTH declare the same zone resources and fight over them on
-// `pulumi up`, and hardening was wrongly coupled to the publish flag.
-// Guarded on the zoneId alone: a prod stack without the config simply skips —
-// hardening activates when prod gains its zone config, independent of
-// publishing. Public repo: no secrets, emails, or IPs — none of these
-// resources need any.
-//
-// Note on the v6 resource names: `ZoneDnssec` (not `ZoneDnssecConfig`) and
-// `ZoneSetting` with `settingId` were both confirmed against the local
-// `@pulumi/cloudflare` v6 typings.
-const hardeningZoneId = stack === "prod" ? config.get("cloudflareZoneId") : undefined;
-if (hardeningZoneId) {
-  const webDomain = config.get("webDomain");
-
-  // DNSSEC signing on the zone. Cloudflare holds the keys and publishes the DS
-  // record; there is deliberately nothing Pulumi-visible beyond the zone.
-  new cloudflare.ZoneDnssec("animichi-dnssec", {
-    zoneId: hardeningZoneId,
-  });
-
-  // CAA: restrict who may issue certificates for the domain to Cloudflare's
-  // Universal SSL certificate partners — letsencrypt.org, pki.goog, ssl.com,
-  // and digicert.com, each encoded as its own record (one tag per record).
-  // Universal SSL keeps working because CF's partner CAs are listed.
-  // The records pin the apex hostname, which prod sets via `webDomain`; until
-  // that config exists they are skipped while DNSSEC, the rate limit, and
-  // HSTS activate on the zoneId alone.
-  if (webDomain !== undefined) {
-    const caaIssuers = [
-      { name: "animichi-caa-letsencrypt", value: "letsencrypt.org" },
-      { name: "animichi-caa-pki-goog", value: "pki.goog; cansignhttpexchanges=yes" },
-      { name: "animichi-caa-ssl-com", value: "ssl.com" },
-      { name: "animichi-caa-digicert", value: "digicert.com; cansignhttpexchanges=yes" },
-    ];
-    for (const { name, value } of caaIssuers) {
-      new cloudflare.DnsRecord(name, {
-        zoneId: hardeningZoneId,
-        name: webDomain,
-        type: "CAA",
-        ttl: 1,
-        data: { flags: 0, tag: "issue", value },
-      });
-    }
-  }
-
-  // One rate-limit rule (the Free plan allows exactly one): a broad
-  // brute-force damper on the API surface — `/v1/*` — keyed by IP + colo,
-  // challenging bursts past 60 requests per 10s for 10s. Turnstile and
-  // per-route quotas do fine-grained control; this only shaves floods.
-  // Scoped to the zone WITHOUT a host match: on this zone only the apex
-  // serves `/v1/*` (www is a placeholder that 301-redirects in
-  // `http_request_dynamic_redirect`, which runs before `http_ratelimit`), so
-  // a bare path match is equivalent to the apex match and needs no `webDomain`
-  // dependency.
-  new cloudflare.Ruleset("animichi-api-rate-limit", {
-    zoneId: hardeningZoneId,
-    name: "zone /v1 rate limit",
-    kind: "zone",
-    phase: "http_ratelimit",
-    description: "Broad rate-limit damper on the /v1 API surface.",
-    rules: [
-      {
-        action: "managed_challenge",
-        expression: `starts_with(http.request.uri.path, "/v1/")`,
-        description: "Challenge bursts past 60 requests per 10s on /v1.",
-        enabled: true,
-        ratelimit: {
-          characteristics: ["ip.src", "cf.colo.id"],
-          period: 10,
-          requestsPerPeriod: 60,
-          mitigationTimeout: 10,
-        },
-      },
-    ],
-  });
-
-  // HSTS on the zone. `include_subdomains` false because the staging subdomain
-  // policy may differ from prod; `preload` false deliberately — preloading is
-  // near-irreversible (removing a domain from the preload list takes months),
-  // so it must be a conscious follow-up decision, not a default.
-  new cloudflare.ZoneSetting("animichi-security-header", {
-    zoneId: hardeningZoneId,
-    settingId: "security_header",
-    value: {
-      strict_transport_security: {
-        enabled: true,
-        max_age: 15552000,
-        include_subdomains: false,
-        preload: false,
-      },
-    },
-  });
-}
-
-// ── Staging: WAF gate ─────────────────────────────────────────────────────────
-// staging runs the same app as production *with anonymous access on*
-// (`ANON_ACCESS_ENABLED = "true"`, workers/edge/wrangler.toml), so there is no login to
-// keep strangers out. A WAF custom rule gates the hostname instead.
-//
-// Why WAF and not Cloudflare Access: Access would do the same job, but the
-// Playwright suite runs against staging and would need a service token. A
-// header is cheaper. Both sit ahead of the Worker either way — custom rules run
-// in `http_request_firewall_custom`, and Cloudflare's own docs are explicit
-// that "Workers runs after the Cloudflare WAF and Cloudflare Access". A blocked
-// request never reaches our code and is not billed as a Worker invocation.
-//
-// Three ways in: an allowlisted source IP (`stagingAllowedIps` — the #769 human
-// path), the legacy `animichi_staging` cookie (set once per browser by hand) or
-// `x-staging-key` header (CI, curl) — #769 card 3 removes the cookie/header
-// path — and the future OIDC exchange endpoint `/staging-gate/exchange` (card 2
-// builds it; passing the WAF is harmless before it exists because unmatched
-// paths 404 at the edge worker). No regex — `matches` is Business+ and this
-// zone is on Free, which allows 5 custom rules.
-//
-// This only works because `workers_dev = false` everywhere (#539): a
-// `*.workers.dev` hostname is not on the zone and would bypass the WAF outright.
-const stagingGateEnabled = config.getBoolean("stagingGateEnabled") ?? false;
-
-// #769: parse the staging allowlist into a Cloudflare `ip.src in { ... }`
-// clause (entries are space-separated inside the braces; plain IPs are allowed
-// alongside CIDRs). Empty list → no clause. Anything else is interpolated into
-// a firewall expression, so a non-IP entry must fail the build loudly.
-export function validateIpEntry(entry: string): boolean {
-  const v4 = entry.match(/^(\d{1,3})\.(\d{1,3})\.(\d{1,3})\.(\d{1,3})(?:\/(\d{1,2}))?$/);
-  if (v4 !== null) {
-    return !v4.slice(1, 5).some((octet) => Number(octet) > 255) && (v4[5] === undefined || Number(v4[5]) <= 32);
-  }
-  const v6 = entry.match(/^[0-9a-fA-F:]+(?:\/(\d{1,3}))?$/);
-  return v6 !== null && (v6[1] === undefined || Number(v6[1]) <= 128);
-}
-
-export function buildIpClause(raw: string): string {
-  const entries = raw.split(",").map((entry) => entry.trim()).filter((entry) => entry.length > 0);
-  if (entries.length === 0) return "";
-  const invalid = entries.find((entry) => !validateIpEntry(entry));
-  if (invalid !== undefined) {
-    throw new Error(`stagingAllowedIps entry "${invalid}" is not a valid IP or CIDR`);
-  }
-  return ` and not (ip.src in {${entries.join(" ")}})`;
-}
-
-// The legacy ruleset identity derives from the zone id, not the list position,
-// so reordering `legacyRedirectZones` never churns Pulumi state. A duplicate
-// id would declare two rulesets with the same Cloudflare `name` in the same
-// zone and fight on `pulumi up`; the list is operator-supplied, so fail the
-// build loudly instead of applying a conflicting declaration.
-export function validateLegacyRedirectZones(zoneIds: string[]): void {
-  const seen = new Set<string>();
-  for (const zoneId of zoneIds) {
-    if (seen.has(zoneId)) {
-      throw new Error(`legacyRedirectZones lists "${zoneId}" more than once`);
-    }
-    seen.add(zoneId);
-  }
-}
-
-// The stack check keeps this resource meaningful only on staging, even if the
-// flag is accidentally enabled on another stack.
-if (stagingGateEnabled && stack === "staging") {
-  const gateZoneId = config.require("cloudflareZoneId");
-  const stagingDomain = config.require("stagingDomain");
-  const gateToken = config.requireSecret("stagingGateToken");
-  const gateTokenValidated = gateToken.apply((t) => {
-    if (!/^[A-Za-z0-9+/=_-]{16,}$/.test(t)) {
-      throw new Error(
-        "stagingGateToken must be >=16 chars of base64/hex/url-safe characters (no quotes, backslashes, or line breaks).",
-      );
-    }
-    return t;
-  });
-
-  // #769: known-human egress IPs (CIDRs, comma-separated). Secret so the list
-  // never lands readable in the public repo or an exported stack backup.
-  const allowedIps = config.getSecret("stagingAllowedIps");
-
-  const ipClause = (allowedIps ?? pulumi.output("")).apply(buildIpClause);
-
-  // `pulumi.interpolate` already propagates secretness from `gateToken`, so the
-  // explicit `pulumi.secret` is belt-and-braces — kept because the cost of
-  // being wrong here is high and asymmetric. Per `AGENTS.md`, every `pulumi up`
-  // is preceded by a `pulumi stack export` copied into R2; a value that is not
-  // marked secret lands in that object in the clear. This repository is public,
-  // so the token must never be reconstructible from anything we publish.
-  // Parenthesised deliberately. Cloudflare's own examples omit them and rely on
-  // `not` binding tighter than `and`, which is correct — but the two failure
-  // modes of getting this wrong are "block every request to staging" and "block
-  // none of them", and neither is visible until the rule is live. Explicit
-  // grouping costs nothing and removes the question.
-  const gateExpression = pulumi.secret(
-    pulumi.interpolate`(http.host eq "${stagingDomain}") and not (http.cookie contains "animichi_staging=${gateTokenValidated}") and not (any(http.request.headers["x-staging-key"][*] eq "${gateTokenValidated}"))${ipClause} and not (http.request.uri.path eq "/staging-gate/exchange")`,
-  );
-
-  new cloudflare.Ruleset("staging-access-gate", {
-    zoneId: gateZoneId,
-    name: "staging access gate",
-    kind: "zone",
-    phase: "http_request_firewall_custom",
-    description: "Restrict the staging hostname to holders of the gate token.",
-    rules: [
-      {
-        action: "block",
-        expression: gateExpression,
-        description: "Block staging traffic without an allowlisted source IP, the gate cookie/header, or the exchange path",
-        enabled: true,
-      },
-    ],
-  });
-}
-
-// ── Staging: per-host config settings (CI smoke) ─────────────────────────────
-// The IP-allowlist gate above is the staging hostname's protection. The WAF's
-// browser-facing defenses would fight it here: a Browser Integrity Check (BIC)
-// or Security Level challenge on the staging host makes the Playwright smoke
-// suite (and any curl) answer a challenge instead of the app. One zone-scoped
-// ruleset turns both off for the staging hostname only — every other hostname
-// on the zone keeps the defaults. Gated on the web routes so preview stacks
-// stay clean: this is meaningful only where the staging hostname actually
-// serves the app.
-if (webRoutesEnabled && stack === "staging") {
-  const settingsZoneId = config.require("cloudflareZoneId");
-  const settingsDomain = config.require("stagingDomain");
-
-  new cloudflare.Ruleset("staging-http-config-settings", {
-    zoneId: settingsZoneId,
-    name: "staging http config settings",
-    kind: "zone",
-    phase: "http_config_settings",
-    description: "Per-host config overrides for the staging hostname.",
-    rules: [
-      {
-        action: "set_config",
-        expression: `http.host eq "${settingsDomain}"`,
-        description: "staging: CI smoke must not be browser-challenged; the IP-allowlist gate owns protection here",
-        enabled: true,
-        actionParameters: {
-          bic: false,
-          securityLevel: "essentially_off",
-        },
-      },
-    ],
-  });
-}
-
-
-export const wave0 = pulumi.output("spike-validated");
-export const catalogBucketName = catalogMediaBucket.name;
-export const tilesBucketName = mapTilesBucket.name;
+export { catalogDatabaseUrl } from "./src/config.ts";
+export { validateLegacyRedirectZones } from "./src/web-routes.ts";
+export { validateIpEntry, buildIpClause } from "./src/staging.ts";
+export {
+  wave0,
+  catalogBucketName,
+  tilesBucketName,
+} from "./src/outputs.ts";
